fix(image): prevent deletion of original file if conversion fails

+++ cn2e-app/src/EventSubscriber/ImageUploadSubscriber.php
@@ -66,7 +66,12 @@ class ImageUploadSubscriber implements EventSubscriberInterface
         $webpPath = $this->converter->convert($originalPath);
 
         // suppression original
-        unlink($originalPath);
+        if (
+            $originalPath !== $webpPath &&
+            file_exists($originalPath)
+        ) {
+            unlink($originalPath);
+        }
 
         $newFilename = basename($webpPath);
